requirements.txt
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1

faker==20.1.0
python-dateutil==2.8.2

psycopg2-binary==2.9.9
sqlalchemy==2.0.23
clickhouse-driver==0.2.6
redis==5.0.1

pyspark==3.4.1

kafka-python==2.0.2
confluent-kafka==2.3.0

minio==7.2.0
boto3==1.34.6

apache-flink==1.17.1

matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

jupyter==1.0.0
jupyterlab==4.0.9
ipywidgets==8.1.1

python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
loguru==0.7.2

pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

requests==2.31.0
tabulate==0.9.0

